src/requirements.txt
requests
pyyaml
pytz
python-dateutil
pandas
noaa_coops
astral
# Development/Testing Dependencies
pytest
requests-mock
